changes for AKPlayer

+++ AudioKit/Common/Internals/AKConverter.swift
@@ -412,7 +412,7 @@ public class AKConverter: NSObject {
         }
         let bufferByteSize: UInt32 = 32_768
         var srcBuffer = [UInt8](repeating: 0, count: 32_768)
-        var sourceFrameOffset: ULONG = 0
+        var sourceFrameOffset: UInt32 = 0
 
         while (true) {
             var fillBufList = AudioBufferList(
